Create compose.yaml for easy local dev setup
I created a docker compose file that encapsulates the database setup for local development.

It spins up 2 containers, one running the API that listens on localhost:5500, and one running postgres database.
The database contianer only listens on the docker network setup by docker compose so it is not accessible outside your local setup.

I updated the README with the docker compose instructions and removed the manual postgres setup.

I needed to update `config.py` to ignore environment variables that aren't used by the `Settings` class. Another way around this would be to make
a separate .env specifically for the database contianer, but I thought this was a bit cleaner.