Update mock_db_session fixture to use postgres container
Instead of using an in-memory sqlite database when running tests, the
mock_db_session fixture now uses a PostgresContainer from the
testcontainers package.

This slows down the startup time of the test suite, but the container
is started up as a 'session' scoped fixture so it only starts once. On
my machine the test suite takes about 10s to run.