enhancement: mock_db_session fixture (#101)
* Update mock_db_session fixture to use postgres container

Instead of using an in-memory sqlite database when running tests, the
mock_db_session fixture now uses a PostgresContainer from the
testcontainers package.

This slows down the startup time of the test suite, but the container
is started up as a 'session' scoped fixture so it only starts once. On
my machine the test suite takes about 10s to run.

* Move fixtures into conftest.py

* Move test_routes.py into new routes directory

* Create fixture that allows tests to run against postgres

Create a new fixure called `override_get_db_session_dependency` that
provides tests with a clean database to test against. It works by
spinning up a postgres container and injecting the connection info into the
get_db_session dependency via the settings object.

Restructure the test suite to mirror the structure of the src
directory.

└── tests
    ├── api
    │   └── routes
    │       ├── test_docker_push_token.py
    │       ├── test_doi.py
    │       ├── test_entrypoint.py
    │       ├── test_garden.py
    │       ├── test_greet.py
    │       ├── test_notebook.py
    │       └── test_routes.py
    ├── auth
    │   └── test_globus_groups.py
    ├── conftest.py
    ├── fixtures
    │   ├── EntrypointCreateRequest-shared-entrypoint.json
    │   ├── EntrypointCreateRequest-with-metadata.json
    │   ├── GardenCreateRequest-shares-entrypoint.json
    │   └── GardenCreateRequest-two-entrypoints.json
    └── __init__.py

* Update ci.yaml

* Add @pytest.mark.container to conditionally run tests

* Rename/restructure stuff in conftest.py for clarity

* Add tests/README.md

* Remove @pytest.mark.integration from non-integration tests

* Handle case when docker is not available

* Automatically skip integrations tests if docker is not available