Add EC2 instance that can communcate with RDS instances
Created terraform resources to provision an EC2 instance with public
IP that can communicate with the RDS instances running in a private
subnet.

This commit adds a new security group that allows incoming connections
on port 22 from the iternet which makes setting up an ssh tunnel to
the RDS instnace possible.

Public ssh keys need to be manually added to the
~/.ssh/authorized_keys file in the running instance in order to
connect via ssh.