Add unit tests verifying failed search index updates are handled correctly
Moved some helper functions to tests/utils.py so they are easier to
use from different test modules.

Made minor tweaks as bugs were found during testing

There are some tests being skipped right now due to some limitations
of this implementation. One is due to not being able to delete an
entrypoint that has an associated garden because the deletes do not
cascade into the 'gardens_entrypoints' table leading to an integrity
issue when deleting the entrypoint.

The others are being skipped because I have not found a good way
of overriding the settings dependency for the 'retry_failed_updates'
loop. Currently the loop is started using the lifespan attribute of
FastAPI, but this is outside of the dependency injection system, so it
is not possible to override in the tests using the same mock_settings
approach as other tests.