Remove 'session' scope from json fixtures
Tests that updated fields in the json fixtures were changing the data
for other tests due to the 'session' scope. Removing the 'session'
scope creates a new instance of the fixture for each test so the data
can be modified within a test without breaking other tests. There is
not a noticeable performance impact on my machine.